rbp_webserver_with_radio/requirements.txt
click==7.1.1
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-crontab>=2.4.1
python-dateutil>=2.8.1
python-vlc==3.0.7110
rpi-rfsniffer==0.1.0
RPi.GPIO==0.7.0
setuptools>=40.8.0
six==1.14.0
Werkzeug>=1.0.1
WTForms==2.2.1